Move templates related utils to a new util class
