+++ apps/console/src/features/applications/utils/application-management-utils.ts
@@ -27,24 +27,18 @@ import startCase from "lodash/startCase";
 import unionBy from "lodash/unionBy";
 import { DocPanelUICardInterface, TechnologyLogos, store } from "../../core";
 import {
-    getApplicationTemplateList,
     getAvailableInboundProtocols,
     getOIDCApplicationConfigurations,
     getSAMLApplicationConfigurations
 } from "../api";
-import { CustomApplicationTemplate } from "../components";
 import { ApplicationManagementConstants } from "../constants";
-import { getDefaultTemplateGroups } from "../meta";
 import {
-    ApplicationTemplateListInterface,
-    ApplicationTemplateListItemInterface,
     AuthProtocolMetaListItemInterface,
     SAMLApplicationConfigurationInterface,
     emptySAMLAppConfiguration
 } from "../models";
 import {
     checkAvailableCustomInboundAuthProtocolMeta,
-    setApplicationTemplates,
     setAvailableCustomInboundAuthProtocolMeta,
     setAvailableInboundAuthProtocolMeta,
     setOIDCApplicationConfigs,
@@ -147,140 +141,6 @@ export class ApplicationManagementUtils {
                 }));
             });
     }
-
-    /**
-     * Retrieve Application template list form the API and sets it in redux state.
-     *
-     * @param {boolean} skipGrouping - Skip grouping of templates.
-     * @return {Promise<void>}
-     */
-    public static getApplicationTemplates = (skipGrouping: boolean = false): Promise<void> => {
-
-        return getApplicationTemplateList()
-            .then((response: ApplicationTemplateListInterface) => {
-
-                const templates: ApplicationTemplateListItemInterface[] = ApplicationManagementUtils
-                    .addCustomTemplates(response.templates, [ CustomApplicationTemplate ], true);
-
-                // Group the templates if `skipGrouping` flag is false.
-                if (!skipGrouping) {
-                    store.dispatch(setApplicationTemplates(ApplicationManagementUtils.groupTemplates(templates), true));
-
-                    return;
-                }
-
-                store.dispatch(setApplicationTemplates(templates));
-            })
-            .catch((error) => {
-                if (error.response && error.response.data && error.response.data.description) {
-                    store.dispatch(addAlert({
-                        description: error.response.data.description,
-                        level: AlertLevels.ERROR,
-                        message: I18n.instance.t("devPortal:components.applications.notifications.fetchTemplates" +
-                            ".error.message")
-                    }));
-
-                    return;
-                }
-
-                store.dispatch(addAlert({
-                    description: I18n.instance.t("devPortal:components.applications.notifications.fetchTemplates" +
-                        ".genericError.description"),
-                    level: AlertLevels.ERROR,
-                    message: I18n.instance.t("devPortal:components.applications.notifications.fetchTemplates" +
-                        ".genericError.message")
-                }));
-            });
-    };
-
-    /**
-     * Sort the application templates based on display order.
-     *
-     * @param {ApplicationTemplateListItemInterface[]} templates - App templates.
-     * @return {ApplicationTemplateListItemInterface[]}
-     */
-    private static sortApplicationTemplates(
-        templates: ApplicationTemplateListItemInterface[]): ApplicationTemplateListItemInterface[] {
-
-        const applicationTemplates = [ ...templates ];
-
-        // Sort templates based  on displayOrder.
-        applicationTemplates.sort(
-            (a, b) =>
-                (a.displayOrder > b.displayOrder) ? 1 : -1);
-
-        return applicationTemplates;
-    }
-
-    /**
-     * Append any custom templates to the existing templates list.
-     *
-     * @param {ApplicationTemplateListItemInterface[]} existingTemplates - Existing templates list.
-     * @param {ApplicationTemplateListItemInterface[]} customTemplates - Set of custom templates to add.
-     * @param {boolean} sort - Should the returning templates be sorted.
-     * @return {ApplicationTemplateListItemInterface[]}
-     */
-    private static addCustomTemplates(existingTemplates: ApplicationTemplateListItemInterface[],
-                                      customTemplates: ApplicationTemplateListItemInterface[],
-                                      sort: boolean = true) {
-
-        if (sort) {
-            return this.sortApplicationTemplates(existingTemplates.concat(customTemplates));
-        }
-
-        return existingTemplates.concat(customTemplates);
-    }
-
-    /**
-     * Group the application templates.
-     *
-     * @param {ApplicationTemplateListItemInterface[]} templates - Application templates.
-     * @return {ApplicationTemplateListItemInterface[]}
-     */
-    private static groupTemplates = (
-        templates: ApplicationTemplateListItemInterface[]): ApplicationTemplateListItemInterface[] => {
-
-        const groupedTemplates: ApplicationTemplateListItemInterface[] = [];
-
-        templates.forEach((template: ApplicationTemplateListItemInterface) => {
-            if (!template.templateGroup) {
-                groupedTemplates.push(template);
-                return;
-            }
-
-            const group = getDefaultTemplateGroups()
-                .find((group) => group.id === template.templateGroup);
-
-            if (!group) {
-                groupedTemplates.push(template);
-                return;
-            }
-
-            if (groupedTemplates.some((groupedTemplate) =>
-                groupedTemplate.id === template.templateGroup)) {
-
-                groupedTemplates.forEach((editingTemplate, index) => {
-                    if (editingTemplate.id === template.templateGroup) {
-                        groupedTemplates[ index ] = {
-                            ...group,
-                            subTemplates: [ ...editingTemplate.subTemplates, template ]
-                        };
-
-                        return;
-                    }
-                });
-
-                return;
-            }
-
-            groupedTemplates.push({
-                ...group,
-                subTemplates: [ template ]
-            });
-        });
-
-        return groupedTemplates;
-    };
 
     /**
      * Build supported technologies list for UI from the given technology types.
