+++ apps/console/src/features/applications/utils/index.ts
@@ -18,3 +18,4 @@
 
 export * from "./adaptive-script-utils";
 export * from "./application-management-utils";
+export * from "./application-template-management-utils";
